requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
django-extensions==3.2.3
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
mysqlclient==2.2.8